Adds node position component with blinking alert
Adds a new component to display node positions on the map,
including details like name, respawn time, and alert status.

Implements a blinking effect and sound alert based on
remaining respawn time milestones to notify the user when a
resource node is about to respawn. Also, provides options to
center the map on the node, reset the timer, delete the node,
and edit the node's properties.

The alert is disabled by default.