Implements pre-start application updater
Adds an updater that checks for updates before the main application starts.

This change introduces a pre-start updater that checks for available updates.
If an update is available, it prompts the user for permission to run the updater.
If the user grants permission, the updater is launched in a detached process,
and the main application exits. This ensures that updates are applied
before the application starts.

The updater is only run if the updater executable exists in the application directory.
The update check is performed by running the updater with the "--check-only" flag.
If the update is available, the updater returns exit code 11.

The updater also includes logic to relaunch the application after the update is complete.

Also fixes an issue where the Updater project would fail to build during publish due to the top-level RID, so explicitly sets RuntimeIdentifier to empty during restore and build of the Updater project.